MPP Updated Rag/Rag 2/hybrid_rerank/requirements.txt
langchain==0.2.14
langchain-openai==0.1.23
langchain-community==0.2.12
chromadb==0.5.5
tiktoken==0.7.0
pypdf==4.3.1
python-dotenv==1.0.1
rank-bm25==0.2.2
